[Bootstrap] removed $('Pokedex-Animate) from every element
$('.Pokedex').click () => $(event.target)
a little bit confused why:

(1) $(event.target) which is instantiated by $('.Pokedex).click
(2) $(event.target)

changed .click(async () => ) to a named function declaration of:
Pokedex.click(async function pokedexClick(event) so that we could have some specifity/targetability with: 
// $(event.target).unbind(pokedexClick)

This problem might be solved by using a useRef which is the preferred way of DOM interactivity. 

I was using useRef for the intersection observer && inView() Hook. You can't use A Hook in a loop so I used jQuery to have some DOM targetability.
Doing so is antipattern/goes-against-unidirectionalDataFlow