[Bootstrap] minimum viability <Button>Component rendered within useRef.map() trick
might be overly personal I am leaning into doing much with bootstrap on this project that arguably shouldn't even be repeated in the first place:
I got a teaching job working with a foundation in a public school

Ill be teaching HTML, CSS, JS, Bootstrap, React,
I actually don't really know bootstrap outside of making a pokemon card. 

I'd like to get a little familiar with it and come up with some gotchas, and tips for the scholars/students.