[MapMaker] unidirectional functionality!!
Problem: 
for example: if you enter 8 into the input: 
8 pokemon from [id# 1 - id#8] render onto the display.

From having 8 pokemon by Id displayed:
If you change the input to 15, an additional 7 pokemon (15 total) Appear Rendered onto our intended display container.

This functionality is not working backwards. 
If you are up to 15 pokemon for example, and change the user-input to '5':

Our 15 pokemon stay rendered onto our container. 
This changing of rendered pokemon based on user input must work both ways for it to be a seamless experience for the user/human