[Bootstrap] somewhat successful import and invocation
[Utility/ClassAction.js]
                   // 2 arguments the element and the className we will be giving it.
export default function ClassAction(elem, action) {
    console.log(elem)
    console.log(action)
    console.log('hey here we go!');
}

[Bootstrap.js]:
👎 Problem:
                          <p onClick={ClassAction()} className="Invisible-P"> P </p>
// this is no good because it runs our classAction function right away,
SO: if we input-enter '10' to see 10 pokemon. Upon load, we get console.log('here we go') (10) ten times without even onClicking the <p> tag

<p onClick={ClassAction(elem, action)} 
// if i want to pass through ClassAction(elem, action) it calls this function automatically as everything is rendered when we setPokedexClick('true')

<p onClick={ClassAction} className="Invisible-P"> P </p>
This is the only arrangement that successfully runs the ClassAction function onClick,
so we see: 1 .log('here we go') for every 1 click onto the <p> tag 

👍 
plan: use this function

ALSO:
it might not hurt to have all the pokeballs have a 1 second time out animation from running the function. 
If they are out of view from intersection observer,

     all we have to do is set a timer:
so the pokeball (in view and out of view) stops animation, then the next to-be-scrolled-into-viewport pokeCard: /img & <pokeBallButton> will be still.    
                        
BONUS: 
to set state to remember which pokeball in the pokeRefs-> Map-Parent index is the one within entry.