[UtilityFunctions/ClassAction] upon Action () => setClass!
This works! 

... dont mean to pass judgment but:
I now see why so many React.js jobs are also node/nearly-full-stack jobs

I'm questioning amongst myself why someone would be into react and shun doing database/back-end work
 when so much of best-practices-reactjs is based on importing/exporting functionality and referenced elements, which is exactly what database/modeling is.

👍 
we dont need those arrays but it was a great idea.
I got the idea from a piece of suggestion:
during the ejs pokedex i had problems with the pokemon being able to sometimes be rendered randomly nearly right next to each other.
The piece of advice was mentioning to store every randomly-appended randomFromPokeAPI-pokemon into an array, and use that array as a catch:

Save the locations appended and don't allow a pokemon to be re-appended.