[Bootstrap]
APIcall() is currently a function that, when invoked, returns all of the pokeapidata.data.results as an iterable object. 

Instead of having 3 exportable utility functions that:
1) return all pokeapi data 
2) return 1 randomized piece of data,
3) return 1 specified by [ pokemon.name || pokemon.id data ]

It'd be experience-boosting to just create:
APIcall() func parameter takes a string-argument allowing for discernment between 3 of the desired outputs associated with making a pokeAPI data call

2 concepts that seem opposed at this point:

The reusability/anti-redundant aspect of having 1 API call function that handles all of our API calls that we would be doing atleast 1 of anyways.

the opposite concept of having 1) allAPIcall() 2) randomPokemonCall() 3) specifyPokemonByNameOrID()
thinking-in-react /  single-responsibility-functions / 1 component for 1 task 

*** i think i see the single-responsibility being the gray area because were still doing the single-task of fetching the data and returning a value.***

project seem already in-the-green on exporting some utility functions 

i already have gap-in-knowledge to bridge on function parameters and creating pure-functions that accept specified input and return output. 

Going with the learn-more option and using APIcall() to accept a 'all' 'random' 'specify' string-argument as first parameter 
we will handle the separation of concerns in this 1 component that does 1 task: accept and return data, but with slight variety.