[PokeAPI] and <RandomPokemonBar.js> loop trick.
APIcall(all||random||specify, 'name||id', pokemontypeForRandomizer, randomAmount)
our random Amount, when specified to be 3 returns 4 random Pokemon. 

        if (typeof randomAmount == 'number' && randomAmount < 5 && randomAmount % 2 !== 0) {            
This is the conditional logic that is accepting our argument.

👎 
1) because of modulo % 2 == 0 / confirm-odd-number-logic: If the goal output is 3 random Pokemon and APIcall(3) returns 4... 
1.1) if I change APIcall(2) it invalidates our modulo logic.

👍 
we can change: 
1) APIcall(2) and then: 
    if (typeof randomAmount == 'number' && randomAmount < 5 && randomAmount + 1  % 2 !== 0) {   
1.1 randomAmount + 1 

1.2: 
    for (i; i < randomAmount -1 ; i++) {        
                RandomAndReset()
            }
we can also change randomAmount - 1 since 3 is returning 4 pokemon.

of course, relative to my experience level, but this is one of the more code-like things i've ever done.