[Evolution] starter-pokemon -> evolveform -> 3rd evolve
Got them linked.
A huge thing I just found out is that I've made about 3-4 or so:
pokemon = {
 name: pokemon.name,
  id: pokemon.id,
  image: githuburl/id
}
i didn't fully know or not know the context/concept of an object being mutable (especially function arguments being mutable once they are in the scope of func)
There should/could be a utility function that exports a pokemon object template for mutating. 
we can use this template over and over. 

There's probably miles to go and i might only be speaking relative to personal XP level.
But it's easy to think: big square page parent container component -----> little navbar components.

the little things like an object template, or creating a bank with different animations to be referenced  by function args, or using a returned Data set from APIcall()

these little things are the bigger surprise than just thinking in state, that might be a misguided assumption / only speaking for experience level.