[randombar] useRef vs createRef
i created a ref.

I had an async mapping issue explained & decided to overcome the async troubles by passing the evolvePokemon([]) state array as props 

i wanted to create a ref() 
because i wanted a relative-to-screen-location hover based fadeIn. 

Also part of my async map problems was i believe i could have fixed it with the componentWILLMount, and was considering switching the class over.
Apparently you can't use [Ref={} the ref prop] unless within a class

The createRef() is to hopefully get a relative element location similar to $(event.target). It would be specified to the specified Ref=[} unfortunately accessible to us in this case as a function component.

instead of createRef i think useRef might work.

***
i also kind of wanted to set this up as a utility function that takes an argument that makes a rendering distinction between:
3 random pokemon rendered in a visual display bar.

or 
3 starter pokemon which would be #1, #4, and #7

im sure theres a way to do beat the map concerns, but to pass the starter pokemon as parent-state props eliminates the ability for us to useState in our exporting function...

console.log(parent)
// this is the reference point that will find our next api-parameter similar to -> element-offsetY-page-location