[Bootstrap] bit of an odd problem:
changing src of image back and forth. 
Img src correctly saves according to corresponding data set by behavior, and does so twice, 

👎 
before the data starts retrieving id # 1 or #4 or #5 upon the 4th or so instance of initiating the behavior.
I also had this set up for state and the changes were being the same type of unpredictable.

Also not a bad idea to consider doing all this changing of image on the 3rd page. 

PAGES:
1) pokeAPI full render (shown here) pokeball.click-> save to fakeDBstate
2) types page. access fakeDBstate/savedPokemon and render them with 2 screens.

3) On the types page: hover over the image of the pokemon on the bottom. Redirects to an information page with displays to navigate pokeAPI data visually.
* we are using map and createRefs() to give us a mapped loop and index from which to access the pokeAPI based on index number with pokemon/${idx#}
* data is intersecting and inconsistent; saving <img src> functionality for a page that doesn't change so much of available index / displayed data (all 151poke)