[Bootstrap]
1) oh wow fixed console.group error it was running over and over in the wild. Snuck it into an onClick event 
2) and isolated the cons.logging() to initiate only upon behavior. 

letting console.group() run over and over again is pretty much the opposite result of the intended outcome of using console.group()
which is to group the logs and keep things non-redundant. 

it would be interesting but probably unneeded to be able to make nicknames for the groups. It's easy to imagine a bigger project and there being use for such a way of organization.
customizing the terminal output to be bold with highlighted color would also be a neat way to specify distinguishment between any 2 groups.