[Bootstrap]
bring in image.
I'm doing so because the pokemon card would of course look better with the pokemon name on it.
if I set the css to be:

<img src={mouseClick === 'false' ? '/infobubble.png' : '' } 
if i do ternary conditional check for any # of mouseclick:
 the info bubble that will take us to our: /pokemon/show/pokemonID   [info page] Will appear for the unclicked pokemon cards.

I was using intersection Observer before as well to allow us to:
1 ) click just under the pokeball on the mouse element.
2) set: {opacity: '0.05'} onto the clicked upon mouse element. (WHILE LEAVING THE OTHER UNCLICKED MOUSE.png to remain {opacity: 1.0}
3)

the re-rendering and re-running every second is causing the setting of variables to be haywire.
while writing this commit i just thought of moving the observer from out of:
function BootStrapScreen() { instantiation 

the observer needs access to the state. It Could help as well to set this up to be its own component.