[Bootstrap] id={ i + 1 } math expressions ok for ID!
interesting: you can perform a mathematical operation to set an ID
1) We are doing so because the data is rendered from a pokeRefs.current/obj.Map() method invocation
2) the .map() method gives our span of data: array-based indexing starting with 0. [bulbasaur. ID #1] shows up as index 0. 
3) {i + 1} is all we have to do to 

also this still works with name.
sometimes intersectionObserver can be funny if you scroll too fast.
it's also already being iffy because were using DOM to change state which is not best practices but seems to be filling gaps in knowledge and experience. 

*** i.e. :
pokemon id# 151 / pokemon.name mew -------> click and the name that pops up is Digglett #50

I went back and tested clicking on 5 pokemon to get the names to pop up and they all popped up.

Instead of accessing and cleaning up the id from the github URL. 
I'm setting an ID from the map render so that we can access 1 clean integer.

*** almost forgot but now relevant: I didn't even know that [className] would be considered a PROP. ***
This is also doing it more from a react way in that we are not stepping outside of our in-line props/params.