[Text] some data/poke !pass length validation
if (text.length. < 15) 

lickitung pops up as " " // empty string (which is what the <ptag> ' '  </p> is set to. 
When class gets removed and the data isn't returned because it invalidates the condition based on [function-Argument/text.length]

This exportable-utility-function and behavior-based-invocation now working as intended.

** can see psychology/habit **
getting a little overly familiar with how peculiar intersection Observer was.

👎  I can see how easy it would be to be frenzied thinking its on the <Bootstrap side> (which up until now it has been)

👍 When the simple fix of [ if (text.length < 15) || if (text.length < 18)] is all you need: look closely without bias.