[Bootstrap]
tucked 2 commits into one. I created an EventTool that allows to describe for an event and return a desired expression.

"To handle the onScroll event in React, set the `onScroll` prop on an element or add an event listener on the `window` object."
were using inline styling and functionality; but nice to see theres a 'window object' way of applying events onto objects.

no need / probably overcomplicate things using such an exported-utility function for EventHandling. relief to have reflex for tooling-function creation.
i dont want to use the EventTool because I want access to event.target 

// {...document.getElementById(i + 1).click()}. <!-- seem to be onto something with this code here  --!>