[Bootstrap]
The problem with intersection Observer was that its running over and over which can be fixed with:
an imported component rather than a utility-function because we need access to the state.

It was clunky to get the pokemon.name to show up because the observer was funny enough that it would work sometimes and sometimes not. 
it would allow the first 3-6 values (bulbasaur/starter-pokemon) but especially scrolling quickly to ID # 50 + There would be mismatched names to image.id

The other alternative I was going to do was to render the:
<div>
<img>
<p tag> mapIndex + 1 </p

render the id onto the pokemon card. if you hover on the pokemon Card then access the id, access Axios, and return the name.
It definitely doesn't look as good as it does if you automatically set the changes invisibly without the user consciously doing things.