[AnimationTool] function() not working
complete function initiates at the last moment and doesn't stick. 
This utility function has been modified to put to use the final argument in the export function parameters. (function() { complete for jq Animation }

👍 in this case the exported myCSS() tool allows us to manipulate the objects. 

👎 it works but it's not sticking / persisting through for some reason.

this is so that any pokemon/data that is saved by clicking the pokeball, Those pokemon that have been saved are {opacity: 0.1} for purpose of indication