[PushAndPop]
str data manipulator that removes and adds /public/energy portion of a url string
this is purposed for adding bg onto an element, and more quickly being able to validate based on existence. 

situation:
[0] useEffect(( )=> and an async IIFE ( () =>() populate state using our modular [bgList()] exported function that holds our url values
[1] Click/Hover/Behavior-Upon (.Circle) -> click circle and it chooses a random bg from the state array that is populated during the actions described above.

[2] 
access state append bg --> now check for the values that aren't the (bg-img-url) value we chose to apply to the bg that was clicked.
create a new state array using every single value, besides the value that was just used. 

* there's an additional state array [tossedValues, setTossedValues] to help double-specify values or not.
using this state array that will hold: (not /img/energy/energyFire.png ---> it holds: energyFire.png)

this tool will allow us to remove and add /img and use the fact that we are doing the same thing over and over to create a reusable tool through which to do so.