[CleanData] tool to parse .nonpretty() 'string'
this export function is specifically to clean up the /apiv2/pokemon/moves the [moves] array that contains all the moves.
if im remembering correctly, there are commas to remove, and there is only 1 array value:
i.e. ()=> moves[0] 'kick, scratch, cut,'
one long string. the moves api is NOT  clean enumerable/iterable data 

this tool will clean that up, but it will also be set up to be low-effort to modify into a function that does different things.

The variety would come from there being several different ways to clean up data coming in, maybe another set has a bunch of ::::: or ///// colons or slashes.

1st argument is array which is the pokemon [moves] or we can check the pokemon on this side and return it.
2) the second argument is the style of the cleaning-up that we will be doing. This example is the main reason why this tool is being created. 

[10:06pm]