[Endpoint]
   // interesting. you can't use string interpolation within concatenation. (as it is its own vehicle of concatenation. still a little surprised.)
            // const newurl = url += `${endpoint}/` += key
            // forgot the slash https://pokeapi.co/api/v2/[ability20] no slash.

this code could be seen as a sports-suicide-drill of just running back and forth for sake of running back and forth. 
👎  I didn't need the CleanUrl tool, but it works!

👍 using the CleanUrl and taking /pokeapi/ability/id and turning it into /id requires us to do a 3 part concatenation using params on other side.
I've never done more than a basic string-to-string concatenation. 

It's Nice to put more weight on the bar so we can do a [3 string .concat()]   || 2 string concat

as new front end instructor:
i'm referencing how easy it is for a new coder to be so confused on concatenation and for what uses and why we would attach 1 string to another


FUN FACT: 
i was surprised to see you be able to use jquery targeting in reacts rendering return statement.
im almost more surprised that you cant use string interpolation in the middle of concatenation. I know interpolation is concatenation but still.

// const newurl = url += `${endpoint}/` += key
this code does not work.



[11:37]