[InfoEndpointBar] ternary operator output
guiding this back to the original way of handling the state & thus container display.
1) click -> change state -> change container display

vs the way i was just trying to set it up.
1) automatically add all of the possible state to the array upon page load through either useEffect(() =>, []}) or I.I.F.E. autoinvoke
2) toggle what is displayed using containerState[0] || containerState[1] and conditionally render elements constrained by state. 

i kind of like this way because state wont be changing and its a different approach to solving a problem that has been used so far. 

*** update ****
props.setEndpointState(movebucket) // this returns values 
bucketBucket[0].push(movebucket) // only returns values when clicked upon. Returns an undefined error if automatically initiated with useEffect or page load.

[4:08pm]