[InfoScreen] add state:
endpointState('') was going to be a specified state but now holds all of the possible data-populations. (explained last commit)

now endpointState([]) holds all of the possible-option values
and there is a new state: 
[endpoint, setEndpoint] 

this new state will be changed and will be the deciding info to base our conditional rendering upon. 
state will be toggled by button click
[5:45pm]