boxingGlove.click->minimumviable /poke/move append
1)  'moves' corresponds to the [pokeapi/v2/pokemon/moves] endpoint.
*** idvalue === 'moves' ? props.setEndpoint(props.endpointState[1][0].move.name) : checkState() ***

no joke no pun this appends the string-data move-API.move.name: 'mega-punch' to the <InfoEndpoint/> element (container)
this element is the parent container above the button bar and is a sibling to this <infoEndpointBar/> code here contained.

I used it before when testing out the <Carousel:
1) 
parentContainer.on('wheel' || 'scroll', () => 
****** props.endpointState[1][scrollState].move.name  

2) increment the scrollState that accepts as the second index which is the counter that loops through the name of all of the pokemons moves.
this can be like a fake scroll without having to loop through all the elements.  



as I type this i love the idea of having all 3 screens be their own component and this move api one having all the moves in one. 
That might be non-performant though given that many pokemon have 100+ moves as this very example does.

[9:00pm]