[Info] length checker [4:40PM]
[1] went digging through the .on(wheel) event object to look for a distinction between upward mousewheelEvents and Downward.
Instead, i've set up a length based way to limit the incrementation. 

Once the limit hits the length it will reset to 0. it would be better to be able to scroll upward or downward through the data but for now as a minimum way to prevent errors, its an ok start

        if (props.moveIndex < movedatalength) { 
// props.endpointState[1][props.moveIndex].