purpose error trying to narrow down event [5:02pm]
purposely creating error with event.clientX. Hoping theres a way to check for negatives and positives to set ++ || -- when setting state when .on('wheel')
it appears most light wheel events are about 1100-1200px so maybe setting around there and making it less sensitive while allowing for upward and downward scrolling through API to give it some realistic touch. 

also a basic state based pokedex. Would be nice to use this simple iteration to model for a typescript version possibly with GraphQL. Would rather be quick. 
And move onto something else but i like the idea of doing this same thing of redoing the project to learn typescript because i wont have to learn about new state and such. One can go about it quicker and learn how to just transfer in already familiar functionality.