[NewRealistic] [7:36pm]
Math.floor(Math.random() * [1, 2, 3].length) // might be trivial but didn't know you could do this.

so there are 15 original types and 7 energy cards.
A few energy cards are going to share types. (psychic type will encompass: 'psychic', 'poison', 'ghost')

the tool is set up to return type in a way of that considers priority of which type we want returned.
i.e--> charizard type[1] === 'flying' type[2] === 'fire' we definitely want it to be fire and not flying. 

next up is looping through [props.fakeDbState] and returning the types so we can