cursor:not-allowed if saved [12:40pm]
if ( poke.saved() --> cursor.notAllowed  )

got this idea while teaching a class.
I was going to get a "kahoot" game set up and noticed that:

[                 ] on the email-address submit-button: if theres no email in the input: 
{ cursor: not-allowed }

I was also explaining that there would be a way to validate email syntax:
if (input.val.includes('@') && input.val.includes(.com)

the kahoot email-address and input based logic did not have the logic set up to reject non-emails. 

and i got the idea of how to structure inputs to scrutinize their values:

 during job applications and seeing some job-apps ask for linkedIn or facebook URL 
and being able to reject the input information if its not a valid facebook url formatting 
(which probably has .includes(facebook.com/) as a conditional argument that must be passed for the input to accept value)

the feature/acknowledgement-of-data-having-been-saved persists because of state and will still change the cursor: none; 

I altered the condition below that so that saving doesn't save if it's already been saved. 
I wanted to unbind the click event, but doing so doesn't work since the logic is glued on as an inline-eventProperty