[Captcha] text changes [ 12:33 am ]
1) the captcha game is based on finding an electric pokemon onto which the user must click to proceed through to the pokedex

2) if (gridItemImage.type !== 'electric') => 
* <p> change p tag text to access type state (which is updated with string-data from the pokemon type)
"That's not an electric pokemon. You clicked a `${typeState||'grass||'fire'||'water'}` Pokemon
** the type is used to create a slightly personalized message that acknowledges the pokemon that was clicked with the change of text 

3) 
let textStart = typeText.text() 

problem is that the purpose of this line of code above is to: 
store the beginning text as a variable so that when the text is changed, there is a stored variable of the original message from which we can access the original value.

👎 
because this is behaviorally-based-function-invoking:
this storing of the variable fires every time a grid item is hovered upon, so it will store the value of the changed text 

the original value gets lost so i created state to save it. there won't be any use of setState('')

** could have also used localStorage but better practice to use state.