[MainScreen]state to parent-prop-as-state [1:32am]
<captcha> -> change [ lock, setLock] to parentState

setLock('unlocked') triggers the falsy block to the rendering ternary operator
If the captcha is completed at all. It shouldn't have to be completed again.

switching to parent state passed down as props instead of using local state allows the data values to persist so the captcha need only to be completed once.