[App]  [11:58am] returned tokens from oAuth2.0 google API
APP.js
<h1> hey </h1>
* firstly i put <h1> into <Routes/> and got a "h1 isn't a valid routes element" error

* I was using the h1 because if you see on the bottom I imported <script google/api/gsi/client/> 
i didn't see that initClient() was a part of the set up to be with rest of init code for 0Auth2.0
function Hello () { .log("hi") }   created at the bottom for minimum-viability. 

                                                                       [ /public/index.html ]
    <script src="https://accounts.google.com/gsi/client" onLoad="initClient()" async defer></script>

    <script>
      var client;
      var access_token;

      function initClient() {
        client = google.accounts.oauth2.initTokenClient({
          client_id: 'YOUR_CLIENT_ID',
          scope: 'https://www.googleapis.com/auth/calendar.readonly \
                  https://www.googleapis.com/auth/contacts.readonly',
          callback: (tokenResponse) => {
            access_token = tokenResponse.access_token;
          },
        });
      }
      function getToken() {
        client.requestAccessToken();
      }
      function revokeToken() {
        google.accounts.oauth2.revoke(access_token, () => {console.log('access token revoked')});
      }
      function loadCalendar() {
        var xhr = new XMLHttpRequest();
        xhr.open('GET', 'https://www.googleapis.com/calendar/v3/calendars/primary/events');
        xhr.setRequestHeader('Authorization', 'Bearer ' + access_token);
        xhr.send();
      }

      function hello() {
        console.log("hey how are you?")
      }
    </script>

separation-of-concerns during 'full-stack' seems very important and it would help to already know 0Auth2.0 before building a log-in 

{accounts: {…}}
{id: {…}, oauth2: {…}}

this is the code that comes back from oAuth 2.0

( lesson spans outwardly-so  because the class project we worked on in bootcamp is already "broken" and the google oAuth2.0 update is the one that did it. 
i went to read the project to get a better understanding of linking [mongo login to react ] front-end-react to back-end-login-credentials and didn't understand how a perfectly good app broke)